+++ ddtrace/contrib/mako/__init__.py
@@ -17,8 +17,12 @@ required_modules = ["mako"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .patch import get_version
-        from .patch import patch
-        from .patch import unpatch
+        # Required to allow users to import from `ddtrace.contrib.mako.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        # Expose public methods
+        from ..internal.mako.patch import get_version
+        from ..internal.mako.patch import patch
+        from ..internal.mako.patch import unpatch
 
         __all__ = ["patch", "unpatch", "get_version"]
