+++ ddtrace/contrib/mariadb/__init__.py
@@ -59,8 +59,12 @@ required_modules = ["mariadb"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .patch import get_version
-        from .patch import patch
-        from .patch import unpatch
+        # Required to allow users to import from `ddtrace.contrib.mariadb.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        # Expose public methods
+        from ..internal.mariadb.patch import get_version
+        from ..internal.mariadb.patch import patch
+        from ..internal.mariadb.patch import unpatch
 
         __all__ = ["patch", "unpatch", "get_version"]
