+++ ddtrace/contrib/mongoengine/__init__.py
@@ -24,7 +24,11 @@ required_modules = ["mongoengine"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .patch import get_version
-        from .patch import patch
+        # Required to allow users to import from `ddtrace.contrib.mongoengine.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        # Expose public methods
+        from ..internal.mongoengine.patch import get_version
+        from ..internal.mongoengine.patch import patch
 
         __all__ = ["patch", "get_version"]
