+++ ddtrace/contrib/mongoengine/patch.py
@@ -1,20 +1,4 @@
-import mongoengine
-
-from .trace import WrappedConnect
+from ..internal.mongoengine.patch import *  # noqa: F401,F403
 
 
-# Original connect function
-_connect = mongoengine.connect
-
-
-def get_version():
-    # type: () -> str
-    return getattr(mongoengine, "__version__", "")
-
-
-def patch():
-    mongoengine.connect = WrappedConnect(_connect)
-
-
-def unpatch():
-    mongoengine.connect = _connect
+# TODO: deprecate and remove this module
