+++ ddtrace/contrib/mysql/__init__.py
@@ -70,7 +70,11 @@ required_modules = ["mysql.connector"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .patch import get_version
-        from .patch import patch
+        # Required to allow users to import from `ddtrace.contrib.mysql.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        # Expose public methods
+        from ..internal.mysql.patch import get_version
+        from ..internal.mysql.patch import patch
 
         __all__ = ["patch", "get_version"]
