+++ ddtrace/contrib/mysqldb/__init__.py
@@ -82,7 +82,11 @@ required_modules = ["MySQLdb"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .patch import get_version
-        from .patch import patch
+        # Required to allow users to import from `ddtrace.contrib.mysqldb.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        # Expose public methods
+        from ..internal.mysqldb.patch import get_version
+        from ..internal.mysqldb.patch import patch
 
         __all__ = ["patch", "get_version"]
