+++ tests/.suitespec.json
@@ -157,6 +157,7 @@
         "mongo": [
             "ddtrace/contrib/pymongo/*",
             "ddtrace/contrib/mongoengine/*",
+            "ddtrace/contrib/internal/mongoengine/*",
             "ddtrace/ext/mongo.py"
         ],
         "pg": [
@@ -297,7 +298,8 @@
             "ddtrace/contrib/rq/*"
         ],
         "mako": [
-            "ddtrace/contrib/mako/*"
+            "ddtrace/contrib/mako/*",
+            "ddtrace/contrib/internal/mako/*"
         ],
         "jinja2": [
             "ddtrace/contrib/jinja2/*",
@@ -343,10 +345,12 @@
             "ddtrace/contrib/internal/httpx/*"
         ],
         "mariadb": [
-            "ddtrace/contrib/mariadb/*"
+            "ddtrace/contrib/mariadb/*",
+            "ddtrace/contrib/internal/mariadb/*"
         ],
         "molten": [
-            "ddtrace/contrib/molten/*"
+            "ddtrace/contrib/molten/*",
+            "ddtrace/contrib/internal/molten/*"
         ],
         "botocore": [
             "ddtrace/contrib/botocore/*",
@@ -358,7 +362,9 @@
         ],
         "mysql": [
             "ddtrace/contrib/mysql/*",
+            "ddtrace/contrib/internal/mysql/*",
             "ddtrace/contrib/mysqldb/*",
+            "ddtrace/contrib/internal/mysqldb/*",
             "ddtrace/contrib/pymysql/*",
             "ddtrace/contrib/aiomysql/*",
             "ddtrace/contrib/internal/aiomysql/*"
